Week13/Day 2/requirements.txt
browser-use
playwright
openai
requests
langchain-openai
langchain-google-genai
python-dotenv
streamlit
nest-asyncio

# then run command 

#playwright install chromium